mastodon-client/requirements.txt
Mastodon.py==1.8.1
python-dotenv==1.0.0
langchain==0.0.153
openai==0.27.5
beautifulsoup4==4.12.2
Flask==2.3.1
types-Flask==1.1.6
sqlalchemy==2.0.12
Flask-SQLAlchemy==3.0.3